# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimInput": {
        "additionalProperties": false,
        "description": "Yêu cầu điều tra thô từ người dùng (API ``POST /api/v1/investigations``).",
        "properties": {
          "claim_text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Claim Text",
            "type": "string"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "dose": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose"
          },
          "drug": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Drug",
            "type": "string"
          },
          "event": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Event",
            "type": "string"
          },
          "population": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "route": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "time_window": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Window"
          }
        },
        "required": [
          "claim_text",
          "drug",
          "event"
        ],
        "title": "ClaimInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MvpApiToken": {
        "description": "Token theo vai trò (investigator/reviewer) gửi qua header X-API-Token.",
        "in": "header",
        "name": "X-API-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Nghiên cứu đối chiếu thuốc trên dữ liệu mô phỏng",
    "title": "MedReview VMEC-03",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/investigations": {
      "get": {
        "description": "Danh sách cuộc điều tra đã lưu (mới nhất trước). Hỗ trợ lọc theo mine_only.",
        "operationId": "list_investigations_api_v1_investigations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mine_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Mine Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "List Investigations",
        "tags": [
          "mvp"
        ]
      },
      "post": {
        "description": "Tạo cuộc điều tra mới và chạy nền; trả 202 + ID.",
        "operationId": "create_investigation_api_v1_investigations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Investigation Api V1 Investigations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Create Investigation",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}": {
      "get": {
        "description": "Trạng thái đầy đủ để frontend polling (counters, gaps, checkpoint).",
        "operationId": "get_investigation_api_v1_investigations__investigation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Investigation Api V1 Investigations  Investigation Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Get Investigation",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/cancel": {
      "post": {
        "description": "Hủy cuộc điều tra đang chạy, đang chờ hoặc chờ reviewer.\n\nChỉ người tạo cuộc điều tra hoặc reviewer mới có quyền hủy.",
        "operationId": "cancel_investigation_api_v1_investigations__investigation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Cancel Investigation",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/continue": {
      "post": {
        "description": "Chạy tiếp sau checkpoint review (giữ nguyên ngân sách đã dùng).\n\nDouble-click được chặn bằng ``Idempotency-Key``: cùng key + cùng version ⇒ không chạy thêm lần nữa.",
        "operationId": "continue_investigation_api_v1_investigations__investigation_id__continue_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ContinueRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Continue Investigation Api V1 Investigations  Investigation Id  Continue Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Continue Investigation",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/documents/{doc_id}": {
      "get": {
        "description": "Nội dung tài liệu gốc + locator để đối chiếu trích dẫn.",
        "operationId": "get_document_api_v1_investigations__investigation_id__documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Document Api V1 Investigations  Investigation Id  Documents  Doc Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/dossier": {
      "get": {
        "description": "Hồ sơ nháp hoặc đã duyệt gần nhất, kèm báo cáo kiểm tra.",
        "operationId": "get_dossier_api_v1_investigations__investigation_id__dossier_get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dossier Api V1 Investigations  Investigation Id  Dossier Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Get Dossier",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/events": {
      "get": {
        "description": "Timeline tiến trình cho polling (``after_id`` để chỉ lấy phần mới).",
        "operationId": "list_events_api_v1_investigations__investigation_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Events Api V1 Investigations  Investigation Id  Events Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/evidence": {
      "get": {
        "description": "Bằng chứng kèm trích dẫn và tài liệu nguồn tương ứng.",
        "operationId": "list_evidence_api_v1_investigations__investigation_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Evidence Api V1 Investigations  Investigation Id  Evidence Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "List Evidence",
        "tags": [
          "mvp"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/export": {
      "get": {
        "description": "Tải Markdown hồ sơ chính thức — chỉ khi có phiên bản đã duyệt.",
        "operationId": "export_dossier_api_v1_investigations__investigation_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Export Dossier",
        "tags": [
          "mvp"
        ]
      }
    },
...
    "/api/v1/investigations/{investigation_id}/trace": {
      "get": {
        "description": "Lấy toàn bộ vết suy luận, các bước chạy LLM và tình trạng ngân sách.",
        "operationId": "get_investigation_trace_api_v1_investigations__investigation_id__trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Investigation Trace Api V1 Investigations  Investigation Id  Trace Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MvpApiToken": []
          }
        ],
        "summary": "Get Investigation Trace",
        "tags": [
          "mvp"
        ]
      }
    },
...
  }
}